// typescript-sdk/src/internal/generated/types/evaluators.generated.ts
export const evaluationResultSchema = z.object({
  status: z.literal("processed"),
  score: z.number().optional(),
  passed: z.boolean().optional(),
  label: z.string().optional(),
  details: z.string().optional(),
  cost: moneySchema.optional(),
  raw_response: z.any().optional(),
});

export const evaluationResultSkippedSchema = z.object({
  status: z.literal("skipped"),
  details: z.string().optional(),
});

export const evaluationResultErrorSchema = z.object({
  status: z.literal("error"),
  error_type: z.string(),
  details: z.string(),
  traceback: z.array(z.string()),
});

export const singleEvaluationResultSchema = z.union([
  evaluationResultSchema,
  evaluationResultSkippedSchema,
  evaluationResultErrorSchema,
]);

export const batchEvaluationResultSchema = z.array(
  singleEvaluationResultSchema,
);

export const evaluatorsSchema = z.object({
  "openai/moderation": z.object({
    settings: z.object({
      model: z
        .union([
          z.literal("text-moderation-stable"),
          z.literal("text-moderation-latest"),
        ])
        .describe(
          "The model version to use, `text-moderation-latest` will be automatically upgraded over time, while `text-moderation-stable` will only be updated with advanced notice by OpenAI.",
        )
        .default("text-moderation-stable"),
      categories: z
        .object({
          harassment: z.boolean().default(true),
          harassment_threatening: z.boolean().default(true),
          hate: z.boolean().default(true),
          hate_threatening: z.boolean().default(true),
          self_harm: z.boolean().default(true),
          self_harm_instructions: z.boolean().default(true),
          self_harm_intent: z.boolean().default(true),
          sexual: z.boolean().default(true),
          sexual_minors: z.boolean().default(true),
          violence: z.boolean().default(true),
          violence_graphic: z.boolean().default(true),
        })
        .describe("The categories of content to check for moderation.")
        .default({
          harassment: true,
          harassment_threatening: true,
          hate: true,
          hate_threatening: true,
          self_harm: true,
          self_harm_instructions: true,
          self_harm_intent: true,
          sexual: true,
          sexual_minors: true,
          violence: true,
          violence_graphic: true,
        }),
    }),
  }),
// ...
});

export type Money = z.infer<typeof moneySchema>;
export type EvaluationResult = z.infer<typeof evaluationResultSchema>;
export type EvaluationResultSkipped = z.infer<
  typeof evaluationResultSkippedSchema
>;
export type EvaluationResultError = z.infer<typeof evaluationResultErrorSchema>;
export type SingleEvaluationResult = z.infer<
  typeof singleEvaluationResultSchema
>;
export type BatchEvaluationResult = z.infer<typeof batchEvaluationResultSchema>;
export type Evaluators = z.infer<typeof evaluatorsSchema>;
export type EvaluatorTypes = keyof Evaluators;